% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_merge_report_column.R
\name{add_merge_report_column}
\alias{add_merge_report_column}
\title{Internal function to add column to merged output table}
\usage{
add_merge_report_column(
  output_table,
  x,
  y,
  by,
  reportvar = getOption("joyn.reportvar")
)
}
\arguments{
\item{output_table}{output data object from joined x and y}

\item{x}{x (or 'left') data object used in the join to create the \code{output_table}}

\item{y}{x (or 'left') data object used in the join to create the \code{output_table}}

\item{by}{character vector: specify \code{by} argument for join,
must be common in all three tables}
}
\value{
data object of same class as \code{output_table} that has additional \verb{_merge} columns -
\verb{_merge} has value 1 if only in \code{x}, 2 if only in \code{y}, and 3 if in both
}
\description{
Internal function to add column to merged output table
}
\keyword{internal}
