% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-joins.R
\name{arguments_checks}
\alias{arguments_checks}
\title{Perform necessary preliminary checks on arguments}
\usage{
arguments_checks(
  x,
  y,
  by,
  copy,
  keep,
  suffix,
  na_matches,
  multiple,
  relationship,
  reportvar
)
}
\value{
list of checked arguments to pass on to the main join function
}
\description{
Perform necessary preliminary checks on arguments
}
\keyword{internal}
