% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_by_vars}
\alias{check_by_vars}
\title{check by input}
\usage{
check_by_vars(by, x, y)
}
\arguments{
\item{by}{a character vector of variables to join by. If NULL, the default,
joyn will do a natural join, using all variables with common names across
the two tables. A message lists the variables so that you can check they're
right (to suppress the message, simply explicitly list the variables that
you want to join). To join by different variables on x and y use a vector
of expressions. For example, \code{by = c("a = b", "z")} will use "a" in x, "b"
in y, and "z" in both tables.}

\item{x}{data frame: referred to \emph{left} in R terminology, or \emph{master} in
Stata terminology.}

\item{y}{data frame: referred to \emph{right} in R terminology, or \emph{using} in
Stata terminology.}
}
\value{
list with information about by variables
}
\description{
check by input
}
\keyword{internal}
