% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-joins.R
\name{check_unmatched_keys}
\alias{check_unmatched_keys}
\title{Conduct all unmatched keys checks and return error if necessary}
\usage{
check_unmatched_keys(x, y, out, by, jn_type)
}
\arguments{
\item{x}{left table}

\item{y}{right table}

\item{out}{output from join}

\item{by}{character vector of keys that x and y are joined by}

\item{jn_type}{character: "left", "right", or "inner"}
}
\value{
error message
}
\description{
Conduct all unmatched keys checks and return error if necessary
}
\keyword{internal}
