% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_var_class}
\alias{check_var_class}
\title{Check join variable class}
\usage{
check_var_class(dt, var)
}
\arguments{
\item{dt}{data.table containing the variable}

\item{var}{Name of the variable to check}
}
\value{
Variable name invisibly if unsupported, otherwise NULL
}
\description{
Checks if a variable in a data.table is of a supported class for joining.
Stores a warning via \code{store_joyn_msg()} if unsupported.
}
\keyword{internal}
