% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_xy}
\alias{check_xy}
\title{check tables X and Y}
\usage{
check_xy(x, y)
}
\arguments{
\item{x}{data frame: referred to \emph{left} in R terminology, or \emph{master} in
Stata terminology.}

\item{y}{data frame: referred to \emph{right} in R terminology, or \emph{using} in
Stata terminology.}
}
\value{
invisible TRUE
}
\description{
This checks are inspired on merge.data.table
}
\keyword{internal}
