% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{correct_names}
\alias{correct_names}
\title{Function used to correct names in input data frames using \code{by} argument}
\usage{
correct_names(by, x, y, order = TRUE)
}
\arguments{
\item{by}{\code{by} argument parsed from higher level function}

\item{x}{left data frame}

\item{y}{right data frame}
}
\value{
list
}
\description{
Function used to correct names in input data frames using \code{by} argument
}
\keyword{internal}
