% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detach_package}
\alias{detach_package}
\title{Detach a package from search list}
\usage{
detach_package(pkg_name)
}
\arguments{
\item{pkg}{name of the package to detach}

\item{character}{logical. TRUE when pkg_name provided as a character string, FALSE otherwise; Default to FALSE}
}
\value{
invisibly detach package from search list
}
\description{
This is an auxiliary function to avoid errors when detaching a package
-e.g., when multiple versions are loaded at the same time
}
\keyword{internal}
