% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_by_vars}
\alias{fix_by_vars}
\title{fix variable names in by argument}
\usage{
fix_by_vars(by, x, y)
}
\arguments{
\item{by}{argument from merge}

\item{x}{left table}

\item{y}{right table}
}
\value{
list
}
\description{
fix variable names in by argument
}
\keyword{internal}
