% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_table.R
\name{freq_table}
\alias{freq_table}
\title{tabulate simple frequencies}
\usage{
freq_table(x, byvar, digits = 1, na.rm = TRUE)
}
\arguments{
\item{x}{data frame}

\item{byvar}{character: name of variable to tabulate. Use Standard evaluation.}

\item{digits}{numeric: number of decimal places to display. Default is 1.}

\item{na.rm}{logical: if TRUE remove NAs from calculations. Default is TRUE}
}
\value{
data.table with frequencies.
}
\description{
tabulate one variable frequencies
}
\examples{
freq_table(x4, "id1")
}
