% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_balanced}
\alias{is_balanced}
\title{Is data frame balanced by group?}
\usage{
is_balanced(df, by, return = c("logic", "table"))
}
\arguments{
\item{df}{data frame}

\item{by}{character: variables to check balance}

\item{return}{character: either "logic" or "table". If logic, returns \code{TRUE}
or \code{FALSE} if data frame is balanced. If "table" returns the unbalanced
observations - i.e. the combinations not found in input \code{df}}
}
\value{
logical if return == "logic", else returns data frame of unbalanced observations
}
\description{
Is data frame balanced by group?
}
\examples{
x1 = data.frame(id = c(1L, 1L, 2L, 3L, NA_integer_),
                t  = c(1L, 2L, 1L, 2L, NA_integer_),
                x  = 11:15)
is_balanced(df = x1,
            by = c("id", "t"),
            return = "table") # returns combo of "id" and "t" not in df
is_balanced(df = x1,
            by = c("id", "t"),
            return = "logic") # returns TRUE or FALSE
}
