% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_id.R
\name{is_id}
\alias{is_id}
\title{Make sure the match type is correct}
\usage{
is_id(dt, by, verbose = TRUE, return_report = FALSE)
}
\arguments{
\item{dt}{either right of left table}

\item{by}{by argument in merge}

\item{verbose}{logical: if TRUE messages will be displayed}

\item{return_report}{logical: if TRUE, returns data with summary of duplicates.
If FALSE, returns logical value depending on whether \code{dt} is uniquely identified
by \code{by}}
}
\value{
logical or data.frame, depending on the value of argument \code{return_report}
}
\description{
Make sure the match type is correct
}
\examples{
is_id(y3, by = "id")
is_id(y3, by = "id", return_report = TRUE)
}
