% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_consistency.R
\name{join_consistency}
\alias{join_consistency}
\title{Make sure the join type is correct}
\usage{
join_consistency(dt, by, dtname = NULL)
}
\arguments{
\item{dt}{either right of left table}

\item{by}{by argument in merge}

\item{dtname}{name of dt for displaying purposes}
}
\value{
TRUE
}
\description{
Make sure the join type is correct
}
