% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joyn-package.R
\docType{package}
\name{joyn-package}
\alias{joyn-package}
\alias{_PACKAGE}
\title{joyn: Tool for Diagnosis of Tables Joins and Complementary Join Features}
\description{
Tool for diagnosing table joins. It combines the speed of `collapse` and `data.table`, the flexibility of `dplyr`, and the diagnosis and features of the `merge` command in `Stata`.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/randrescastaneda/joyn}
  \item \url{https://rossanatat.github.io/joyn/}
  \item Report bugs at \url{https://github.com/randrescastaneda/joyn/issues}
}

}
\author{
\strong{Maintainer}: R.Andres Castaneda \email{acastanedaa@worldbank.org}

Authors:
\itemize{
  \item Zander Prinsloo \email{zprinsloo@worldbank.org}
}

Other contributors:
\itemize{
  \item Rossana Tatulli \email{rtatulli@worldbank.org} [contributor]
}

}
\keyword{internal}
