% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_display.R
\name{joyn_msg}
\alias{joyn_msg}
\title{display type of joyn message}
\usage{
joyn_msg(type = c("all", type_choices()))
}
\arguments{
\item{type}{character: one or more of the following:
"info", "note", "warn", "timing", or "err" or \code{all}}
}
\value{
returns data frame with message invisibly. print message in console
}
\description{
display type of joyn message
}
\examples{
# Storing msg with type "info"
joyn:::store_msg("info",
  ok = cli::symbol$tick, "  ",
  pale = "This is an info message")

# Storing msg with type "warn"
joyn:::store_msg(
  type        = "warn",
  warn        = paste(cli::symbol$warn, "\nWarning:"),
  pale        = "  joyn does not currently allow inequality joins, so",
  bolded_pale = "  keep = NULL",
  pale        = "  will retain only keys in x. Equivalent to `keep = FALSE`"
)

joyn_msg("all")
}
