% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_display.R
\name{joyn_msgs_exist}
\alias{joyn_msgs_exist}
\title{Presence of joyn msgs in the environment}
\usage{
joyn_msgs_exist()
}
\value{
invisible TRUE
}
\description{
Checks the presence of joyn messages stored in joyn environment
}
\examples{
\dontrun{
Storing a message
store_msg("info", "simple message")
Checking if it exists in the environment
print(joyn_msgs_exist())
}
}
