% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joyn_workhorse.R
\name{joyn_workhorse}
\alias{joyn_workhorse}
\title{Internal workhorse join function, used in the backend of \code{joyn}}
\usage{
joyn_workhorse(
  x,
  y,
  by = intersect(names(x), names(y)),
  match_type = c("1:1", "1:m", "m:1", "m:m"),
  suffix = getOption("joyn.suffixes")
)
}
\arguments{
\item{x}{data object, "left" or "master"}

\item{y}{data object, "right" or "using"}

\item{by}{atomic character vector: key specifying join}

\item{match_type}{atomic character vector length 1: either "1:1" (default)
"1:m", "m:1", or "m:m". If "m:m" then executes \code{data.table::merge.data.table}
in the backend, otherwise uses \code{collapse::join()}}

\item{suffix}{atomic character vector: give suffix to columns common to both
\code{x} and \code{y}}

\item{keep}{atomic character vector of length 1:
One of \emph{"full"}, \emph{"left"}, \emph{"master"}, \emph{"right"},
\emph{"using"}, \emph{"inner"}. Default is \emph{"full"}. Even though this is not the
regular behavior of joins in R, the objective of \code{joyn} is to present a
diagnosis of the join which requires a full join. That is why the default
is a a full join. Yet, if \emph{"left"} or \emph{"master"}, it keeps the observations
that matched in both tables and the ones that did not match in x. The ones
in y will be discarded. If \emph{"right"} or \emph{"using"}, it keeps the observations
that matched in both tables and the ones that did not match in y. The ones in x
will be discarded. If \emph{"inner"}, it only keeps the observations that
matched both tables.}
}
\description{
Always executes a full join. Gives comm
}
