% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_ids.R
\name{possible_ids_old}
\alias{possible_ids_old}
\title{Find possible unique identifies of data frame}
\usage{
possible_ids_old(
  dt,
  exclude = NULL,
  include = NULL,
  verbose = getOption("possible_ids.verbose")
)
}
\arguments{
\item{dt}{data frame}

\item{exclude}{character: Exclude variables to be selected as identifiers. It
could be either the name of the variables of one type of the variable
prefixed by "_". For instance, "_numeric" or "_character".}

\item{include}{character: Name of variable to be included, that might belong
to the group excluded in the \code{exclude}}

\item{verbose}{logical: If FALSE no message will be displayed. Default is
TRUE}
}
\value{
list with possible identifiers
}
\description{
Identify possible variables uniquely identifying x
}
\examples{
library(data.table)
x4 = data.table(id1 = c(1, 1, 2, 3, 3),
                id2 = c(1, 1, 2, 3, 4),
                t   = c(1L, 2L, 1L, 2L, NA_integer_),
                x   = c(16, 12, NA, NA, 15))
possible_ids_old(x4)
}
\keyword{internal}
