% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-joins.R
\name{process_by_vector}
\alias{process_by_vector}
\title{Process the \code{by} vector}
\usage{
process_by_vector(by, input = c("left", "right"))
}
\arguments{
\item{by}{character vector: by argument for join}

\item{input}{character: either "left" or "right", indicating
whether to give the left or right side of the equals ("=") if
the equals is part of the \code{by} vector}
}
\value{
character vector
}
\description{
Gives as output a vector of names to be used for the specified
table that correspond to the \code{by} argument for that table
}
\examples{
process_by_vector(by = c("An = foo", "example"), input = "left")
}
\keyword{internal}
