% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_table.R
\name{report_from_attr}
\alias{report_from_attr}
\title{Report frequencies from attributes in report var}
\usage{
report_from_attr(x, y, reportvar)
}
\arguments{
\item{x}{dataframe from \link{joyn_workhorse}}

\item{y}{dataframe from original merge ("right" or "using")}
}
\value{
dataframe with frequencies of report var
}
\description{
Report frequencies from attributes in report var
}
\keyword{internal}
