% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-joins.R
\name{set_col_names}
\alias{set_col_names}
\title{Add x key var and y key var (with suffixes) to x and y
-when joining by different variables and keep is true}
\usage{
set_col_names(x, y, by, suffix, jn_type)
}
\arguments{
\item{x}{data table: left table}

\item{y}{data table: right table}

\item{by}{character vector of variables to join by}

\item{suffix}{character(2) specifying the suffixes to be used for making non-by column names unique}

\item{jn_type}{character specifying type of join}
}
\value{
list containing x and y
}
\description{
Add x key var and y key var (with suffixes) to x and y
-when joining by different variables and keep is true
}
\keyword{internal}
