% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{set_joyn_options}
\alias{set_joyn_options}
\title{Set joyn options}
\usage{
set_joyn_options(..., env = .joynenv)
}
\arguments{
\item{...}{pairs of option = value}

\item{env}{environment, which is joyn environment by default}
}
\value{
joyn new options and values invisibly as a list
}
\description{
This function is used to change the value of one or more joyn options
}
\examples{
joyn:::set_joyn_options(joyn.verbose = FALSE, joyn.reportvar = "joyn_status")
joyn:::set_joyn_options() # return to default options
}
