% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_display.R
\name{store_msg}
\alias{store_msg}
\title{Store joyn message to .joynenv environment}
\usage{
store_msg(type, ...)
}
\arguments{
\item{type}{character: type of message}

\item{...}{combination of type and text in the form \verb{type1 = text1, type2 = text2}, etc.}
}
\value{
current message data frame invisibly
}
\description{
Store joyn message to .joynenv environment
}
\keyword{internal}
