% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_display.R
\name{style}
\alias{style}
\title{style of text displayed}
\usage{
style(..., sep = "")
}
\arguments{
\item{...}{combination of type and text in the form
\verb{type1 = text1, type2 = text2}}

\item{sep}{a character string to separate the terms to \link{paste}}
}
\value{
formatted text
}
\description{
This is an adaptation from
https://github.com/r-lib/pkgbuild/blob/3ba537ab8a6ac07d3fe11c17543677d2a0786be6/R/styles.R
}
\seealso{
Messages functions
\code{\link{clear_joynenv}()},
\code{\link{joyn_msg}()},
\code{\link{joyn_msgs_exist}()},
\code{\link{joyn_report}()},
\code{\link{msg_type_dt}()},
\code{\link{store_msg}()},
\code{\link{type_choices}()}
}
\concept{messages}
\keyword{internal}
