% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_display.R
\name{type_choices}
\alias{type_choices}
\title{Choice of messages}
\usage{
type_choices()
}
\value{
character vector with choices of types
}
\description{
Choice of messages
}
\seealso{
Messages functions
\code{\link{clear_joynenv}()},
\code{\link{joyn_msg}()},
\code{\link{joyn_msgs_exist}()},
\code{\link{joyn_report}()},
\code{\link{msg_type_dt}()},
\code{\link{store_msg}()},
\code{\link{style}()}
}
\concept{messages}
\keyword{internal}
