% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unmask_function_curr_env}
\alias{unmask_function_curr_env}
\title{Unmask joyn's functions}
\usage{
unmask_function_curr_env(fun_name, pkg_name)
}
\arguments{
\item{fun_name}{character vector of one or more functions to unmask}

\item{pkg_name}{character specifying package from which joyn masks the function(s)}
}
\value{
invisible TRUE

invisible TRUE
}
\description{
Unmask joyn's functions
}
\keyword{-to}
\keyword{Unmask}
\keyword{be}
\keyword{current}
\keyword{done}
\keyword{environment}
\keyword{export}
\keyword{function(s)}
\keyword{in}
\keyword{internal}
\keyword{joyn's}
\keyword{once}
\keyword{replaced}
\keyword{with}
