% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unmask_joyn_fun_ns}
\alias{unmask_joyn_fun_ns}
\title{Unmask joyn function (namespace version)}
\usage{
unmask_joyn_fun_ns(fun_name, pkg_name)
}
\arguments{
\item{fun_name}{character of function to unmask}

\item{pkg_name}{character specifying package from which joyn masks the function}
}
\value{
invisibly unmask function that joyn masks
}
\description{
Unmask joyn function (namespace version)
}
\keyword{internal}
