% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unmask_joyn_fun_old}
\alias{unmask_joyn_fun_old}
\title{Unmask joyn's functions}
\usage{
unmask_joyn_fun_old(fun_name, pkg_name)
}
\arguments{
\item{fun_name}{character vector of one or more functions to unmask}

\item{pkg_name}{character specifying package from which joyn masks the function(s)}
}
\value{
invisible TRUE
}
\description{
Unmask joyn's functions
}
\examples{
unmask_joyn_fun(fun_name = c("left_join", "right_join", "full_join"),
                pkg_name = "dplyr")

unmask_joyn_fun(fun_name = "inner_join",
                pkg_name = "dplyr")

}
