% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-joins.R
\name{unmatched_keys}
\alias{unmatched_keys}
\title{Detect unmatched keys}
\usage{
unmatched_keys(x, y, by, output, jn_type)
}
\arguments{
\item{x}{data frame: left table}

\item{y}{data frame: right table}

\item{by}{character vector of variables to join by}

\item{output}{joined data table, output of either a left, right or inner join}

\item{jn_type}{character specifying type of join}
}
\value{
logical TRUE if unmatched keys are found, FALSE if no unmacthed keys are found
}
\description{
Detect unmatched keys
}
\keyword{internal}
