% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_NAs.R
\name{update_NAs}
\alias{update_NAs}
\title{Update NA values}
\usage{
update_NAs(dt, var)
}
\arguments{
\item{dt}{data.table}

\item{var}{variable to be updated}
}
\value{
data.table
}
\description{
Update NA values
}
