% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_values.R
\name{update_values}
\alias{update_values}
\title{update values in x with data from y}
\usage{
update_values(dt, var)
}
\arguments{
\item{dt}{joined table}

\item{var}{variable to be updated}
}
\value{
data.table
}
\description{
update values in x with data from y
}
