% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_by_vars}
\alias{check_by_vars}
\title{check by input}
\usage{
check_by_vars(by, x, y)
}
\arguments{
\item{by}{A vector of shared column names in \code{x} and \code{y} to merge on.
This defaults to the shared key columns between the two tables.
If \code{y} has no key columns, this defaults to the key of \code{x}.}

\item{x, y}{\code{data table}s. \code{y} is coerced to a \code{data.table} if
it isn't one already.}
}
\value{
list with information about by variables
}
\description{
check by input
}
\keyword{internal}
