% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_match_type}
\alias{check_match_type}
\title{check match type consistency}
\usage{
check_match_type(x, y, by, match_type, verbose)
}
\arguments{
\item{x, y}{\code{data table}s. \code{y} is coerced to a \code{data.table} if
it isn't one already.}

\item{by}{A vector of shared column names in \code{x} and \code{y} to merge on.
This defaults to the shared key columns between the two tables.
If \code{y} has no key columns, this defaults to the key of \code{x}.}

\item{match_type}{character: one of \emph{"m:m"}, \emph{"m:1"}, \emph{"1:m"}, \emph{"1:1"}.
Default is \emph{"1:1"} since this the most restrictive. However, following
Stata's recommendation, it is better to be explicit and use any of the
other three match types (See details in \emph{match types sections}).}
}
\value{
character vector from \link{split_match_type}
}
\description{
check match type consistency
}
\keyword{internal}
