% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_new_y_vars}
\alias{check_new_y_vars}
\title{rename vars in y so they are different to x's when joined}
\usage{
check_new_y_vars(x, by, y_vars_to_keep)
}
\arguments{
\item{x}{master table}

\item{by}{character: by vars}

\item{y_vars_to_keep}{character vector of variables to keep}
}
\value{
vector with new variable names for y
}
\description{
rename vars in y so they are different to x's when joined
}
\keyword{internal}
