% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_reportvar}
\alias{check_reportvar}
\title{Check reportvar input}
\usage{
check_reportvar(reportvar, verbose = getOption("joyn.verbose"))
}
\value{
if input reportvar is character, return valid name for the report var. If NULL or FALSE, return NULL.
}
\description{
check reporting variable (storing joyn's report).
}
\examples{
# When null - reporting variable not returned in merged dt
check_reportvar(reportvar = NULL)
# When FALSE - reporting variable not returned in merged dt
check_reportvar(reportvar = FALSE)
# When character
check_reportvar(reportvar = ".joyn")
}
\keyword{internal}
