% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_reportvar}
\alias{check_reportvar}
\title{Check reporting variable}
\usage{
check_reportvar(reportvar, verbose = getOption("joyn.verbose"))
}
\value{
if input reportvar is character, return valid name for the report var. If NULL or FALSE, return NULL.
}
\description{
check reportvar input \if{html}{\out{<br>}} If resulting data frame has a reporting variable (storing joyn's report), check and return a valid name.
}
\examples{
\dontrun{
# When null - reporting variable not returned in merged dt
joyn:::check_reportvar(reportvar = NULL)
# When FALSE - reporting variable not returned in merged dt
joyn:::check_reportvar(reportvar = FALSE)
# When character
joyn:::check_reportvar(reportvar = ".joyn")
}
}
\keyword{internal}
