% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_reportvar}
\alias{check_reportvar}
\title{check reportvar input}
\usage{
check_reportvar(reportvar, verbose = getOption("joyn.verbose"))
}
\value{
if character, it returns valid name. If NULL or FALSE, returns NULL.
}
\description{
check reportvar input
}
\keyword{internal}
