% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_y_vars_to_keep}
\alias{check_y_vars_to_keep}
\title{check variables in Y that will be kept in returning table}
\usage{
check_y_vars_to_keep(y_vars_to_keep, y, by)
}
\arguments{
\item{by}{A vector of shared column names in \code{x} and \code{y} to merge on.
This defaults to the shared key columns between the two tables.
If \code{y} has no key columns, this defaults to the key of \code{x}.}
}
\value{
character vector with variable names from Y table
}
\description{
check variables in Y that will be kept in returning table
}
\keyword{internal}
