% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_y_vars_to_keep}
\alias{check_y_vars_to_keep}
\title{check variables in Y that will be kept in returning table}
\usage{
check_y_vars_to_keep(y_vars_to_keep, y, by)
}
\arguments{
\item{y_vars_to_keep}{character: Vector of variable names that will be kept after the
merge. If TRUE (the default), it keeps all the brings all the variables in
y into x. If FALSE or NULL, it does not bring any variable into x, but a
report will be generated.}

\item{y}{data frame: referred to \emph{right} in R terminology, or \emph{using} in
Stata terminology.}

\item{by}{a character vector of variables to join by. If NULL, the default,
joyn will do a natural join, using all variables with common names across
the two tables. A message lists the variables so that you can check they're
right (to suppress the message, simply explicitly list the variables that
you want to join). To join by different variables on x and y use a vector
of expressions. For example, \code{by = c("a = b", "z")} will use "a" in x, "b"
in y, and "z" in both tables.}
}
\value{
charactere vector with variable names from Y table
}
\description{
check variables in Y that will be kept in returning table
}
\keyword{internal}
