% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_table.R
\name{freq_table}
\alias{freq_table}
\title{tabulate simple frequencies}
\usage{
freq_table(x, byvar, digits = 1)
}
\arguments{
\item{x}{data frame}

\item{byvar}{character: name of variable to tabulate. Use STandard evaluation.}

\item{digits}{numeric: number of decimal places to display. Default is 1.}
}
\value{

}
\description{
tabulate one variable in data frame to substitute base::table and
janitor::tabyl using data.table
}
\examples{
freq_table(x4, "id1")
}
