% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_match_type_error}
\alias{is_match_type_error}
\title{confirm if match_type_error}
\usage{
is_match_type_error(x, name, by, verbose, match_type_error)
}
\arguments{
\item{name}{name of variable}

\item{by}{A vector of shared column names in \code{x} and \code{y} to merge on.
This defaults to the shared key columns between the two tables.
If \code{y} has no key columns, this defaults to the key of \code{x}.}

\item{match_type_error}{logical: from existing code}
}
\value{
logical
}
\description{
confirm if match_type_error
}
\examples{
# example with match type error
TODO
}
\keyword{internal}
