% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_valid_m_key}
\alias{is_valid_m_key}
\title{Check whether specified "many" relationship is valid}
\usage{
is_valid_m_key(dt, by)
}
\arguments{
\item{dt}{data object}

\item{by}{character vector: specified keys, already fixed}
}
\value{
logical: \code{TRUE} if valid, \code{FALSE} if uniquely identified
}
\description{
Check whether specified "many" relationship is valid
}
\keyword{internal}
