% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_display.R
\name{joyn_msg}
\alias{joyn_msg}
\title{add joyn message to .joynenv environment}
\usage{
joyn_msg(type, ...)
}
\arguments{
\item{type}{character: type of message}

\item{...}{combination of type and text in the form \verb{type1 = text1, type2 = text2}, etc.}
}
\value{
invisible TRUE
}
\description{
add joyn message to .joynenv environment
}
\keyword{internal}
