% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_display.R
\name{joyn_msg}
\alias{joyn_msg}
\title{display type of joyn message}
\usage{
joyn_msg(type = c("all", "info", "note", "warn", "timing"))
}
\arguments{
\item{type}{character: one or more of \code{c("all", "info", "note", "warn", "timing")}}
}
\value{
returns data frame with message invisibly. print message in console
}
\description{
display type of joyn message
}
\examples{
joyn:::store_msg("info", ok = cli::symbol$tick, "  ", pale = "This is an info message")
joyn:::store_msg("warn", err = cli::symbol$cross, "  ", note = "This is a warning message")
joyn_msg("all")
}
