% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_display.R
\name{joyn_msg}
\alias{joyn_msg}
\title{display type of joyn message}
\usage{
joyn_msg(type = c("all", type_choices()), msg = NULL)
}
\arguments{
\item{type}{character: one or more of the following: info, note, warn, timing, err
cli::format_inline("{.or {.val {type_choices()}}}")\code{or}all`}

\item{msg}{character vector to be parsed to \code{\link[cli:cli_abort]{cli::cli_abort()}}. Default is
NULL. It only works if \code{"err" \%in\% type}}
}
\value{
returns data frame with message invisibly. print message in console
}
\description{
display type of joyn message
}
\examples{
# Storing msg with type "info"
joyn:::store_msg("info",
  ok = cli::symbol$tick, "  ",
  pale = "This is an info message")

# Storing msg with type "warn"
joyn:::store_msg("warn",
  err = cli::symbol$cross, "  ",
  note = "This is a warning message")

joyn_msg("all")
}
