% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_display.R
\name{joyn_msg}
\alias{joyn_msg}
\title{display type of joyn message}
\usage{
joyn_msg(msg_type = getOption("joyn.msg_type"), msg = NULL)
}
\arguments{
\item{msg_type}{character: one or more of the following:
all, basic, info, note, warn, timing, or err}

\item{msg}{character vector to be parsed to \code{\link[cli:cli_abort]{cli::cli_abort()}}. Default is
NULL. It only works if \code{"err" \%in\% msg_type}}
}
\value{
returns data frame with message invisibly. print message in console
}
\description{
display type of joyn message
}
\examples{
# Storing msg with msg_type "info"
joyn:::store_msg("info",
  ok = cli::symbol$tick, "  ",
  pale = "This is an info message")

# Storing msg with msg_type "warn"
joyn:::store_msg("warn",
  err = cli::symbol$cross, "  ",
  note = "This is a warning message")

joyn_msg("all")
}
