% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge}
\alias{merge}
\title{Merge two data.frames}
\usage{
merge(
  x,
  y,
  by = NULL,
  byexp = NULL,
  roll = NULL,
  yvars = NULL,
  type = c("m:m", "m:1", "1:m", "1:1"),
  keep = c("both", "full", "left", "master", "right", "using", "inner"),
  update_values = FALSE,
  updateNA = update_values,
  reportvar = "report",
  reporttype = c("character", "numeric"),
  keep_y_in_x = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{x}{data frame: In Stata terminology, it is known as the master data}

\item{y}{data frame: In Stata terminology, it is known as the using data}

\item{byexp}{}
}
\value{

}
\description{
Merge two data.frames
}
