% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_display.R
\name{msg_type_dt}
\alias{msg_type_dt}
\title{convert style of joyn message to data frame containing type and message}
\usage{
msg_type_dt(type, ...)
}
\value{
data frame with two variables, type and msg
}
\description{
convert style of joyn message to data frame containing type and message
}
\seealso{
Messages functions
\code{\link{clear_joynenv}()},
\code{\link{joyn_msg}()},
\code{\link{joyn_msgs_exist}()},
\code{\link{joyn_report}()},
\code{\link{store_msg}()},
\code{\link{style}()},
\code{\link{type_choices}()}
}
\concept{messages}
\keyword{internal}
