% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_display.R
\name{msg_type_dt}
\alias{msg_type_dt}
\title{convert style to data frame}
\usage{
msg_type_dt(type, ...)
}
\arguments{
\item{type}{character: one or more of \code{c("all", "info", "note", "warn")}}
}
\value{
data frame
}
\description{
convert style to data frame
}
\keyword{internal}
