% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_display.R
\name{msg_type_dt}
\alias{msg_type_dt}
\title{convert style of joyn message to data frame containing type and message}
\usage{
msg_type_dt(type, ...)
}
\arguments{
\item{type}{character: one or more of the following:
"info", "note", "warn", "timing", or "err" or \code{all}}
}
\value{
data frame with two variables, type and msg
}
\description{
convert style of joyn message to data frame containing type and message
}
\keyword{internal}
