% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_ids.R
\name{possible_ids}
\alias{possible_ids}
\title{Find possible unique identifies of data frame}
\usage{
possible_ids(dt, verbose = TRUE, exclude = NULL, include = NULL)
}
\arguments{
\item{dt}{data frame}

\item{verbose}{logical: fi FALSE no message will be displayed. Default is
TRUE}

\item{exclude}{character: Exclude variables to be selected as identifiers. It
could be either the name of the variables of one type of the variable
prefixed by "_". For instance, "_numeric" or "_character".}

\item{include}{character: Name of variable to be included, that might belong
to the group excluded in the \code{exclude}}
}
\value{
list with possible identifiers
}
\description{
Find possible unique identifies of data frame
}
\examples{
possible_ids(x4)
}
