% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_display.R
\name{store_joyn_msg}
\alias{store_joyn_msg}
\title{Wrapper for store_msg function
This function serves as a wrapper for the store_msg function, which is used to store various types of messages within the .joyn environment.
:errors, warnings, timing information, or info}
\usage{
store_joyn_msg(err = NULL, warn = NULL, timing = NULL, info = NULL)
}
\arguments{
\item{err}{A character string representing an error message to be stored. Default value is NULL}

\item{warn}{A character string representing a warning message to be stored. Default value is NULL}

\item{timing}{A character string representing a timing message to be stored. Default value is NULL}

\item{info}{A character string representing an info message to be stored. Default value is NULL}
}
\value{
invisible TRUE
}
\description{
Wrapper for store_msg function
This function serves as a wrapper for the store_msg function, which is used to store various types of messages within the .joyn environment.
:errors, warnings, timing information, or info
}
\section{Hot to pass the message string}{

The function allows for the customization of the message string using {cli} classes to emphasize specific components of the message
Here's how to format the message string:
*For variables:            .strongVar   --example: "{.strongVar {reportvar}}"
*For function arguments:   .strongArg   --example: "{.strongArg {y_vars_to_keep}}"
*For dt/df:                .strongTable --example: "{.strongTable x}"
*For text/anything else:   .strong      --example: "reportvar is {.strong NOT} returned"
}

\keyword{internal}
