% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_display.R
\name{store_msg}
\alias{store_msg}
\title{Store joyn message to .joynenv environment}
\usage{
store_msg(type, ...)
}
\arguments{
\item{...}{combination of type and text in the form \verb{style1 = text1, style2 = text2}, etc.}
}
\value{
current message data frame invisibly
}
\description{
Store joyn message to .joynenv environment
}
\examples{
# Storing msg with msg_type "info"
joyn:::store_msg("info",
  ok = cli::symbol$tick, "  ",
  pale = "This is an info message")

# Storing msg with msg_type "warn"
joyn:::store_msg("warn",
  err = cli::symbol$cross, "  ",
  note = "This is a warning message")
}
\seealso{
Messages functions
\code{\link{clear_joynenv}()},
\code{\link{joyn_msgs_exist}()},
\code{\link{joyn_msg}()},
\code{\link{joyn_report}()},
\code{\link{msg_type_dt}()},
\code{\link{style}()},
\code{\link{type_choices}()}
}
\concept{messages}
\keyword{internal}
