% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_display.R
\name{store_msg}
\alias{store_msg}
\title{Store joyn message to .joynenv environment}
\usage{
store_msg(type, ...)
}
\arguments{
\item{type}{character: one or more of the following:
"info", "note", "warn", "timing", or "err"}

\item{...}{combination of type and text in the form \verb{style1 = text1, style2 = text2}, etc.}
}
\value{
current message data frame invisibly
}
\description{
Store joyn message to .joynenv environment
}
\examples{
# With type = info and text = "simple message"

store_msg("info", "simple message")
rlang::env_has(env = .joynenv, "joyn_msgs")

}
\keyword{internal}
