% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_display.R
\name{style}
\alias{style}
\title{style of text displayed}
\usage{
style(..., sep = "")
}
\arguments{
\item{...}{combination of type and text in the form
\verb{type1 = text1, type2 = text2}}

\item{sep}{a character string to separate the terms to \link{paste}}
}
\value{
formated text
}
\description{
This is an adaptation from
https://github.com/r-lib/pkgbuild/blob/3ba537ab8a6ac07d3fe11c17543677d2a0786be6/R/styles.R
}
\keyword{internal}
