% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unmask_joyn}
\alias{unmask_joyn}
\title{Unmask joyn functions}
\usage{
unmask_joyn(fun_name, pkg_name = "dplyr")
}
\arguments{
\item{fun_name}{character vector of one or more functions to unmask}

\item{pkg_name}{character specifying package from which joyn masks the function(s)}
}
\value{
invisible list
}
\description{
Unmask joyn functions
}
