% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unmask_joyn_fun}
\alias{unmask_joyn_fun}
\title{Unmask joyn's functions}
\usage{
unmask_joyn_fun(fun_name, pkg_name)
}
\arguments{
\item{fun_name}{character vector of one or more functions to unmask}

\item{pkg_name}{character specifying package from which joyn masks the function(s)}
}
\value{
invisible TRUE

invisible TRUE
}
\description{
Unmask joyn's functions
}
\keyword{-this}
\keyword{-to}
\keyword{Unmask}
\keyword{be}
\keyword{done}
\keyword{export}
\keyword{function(s)}
\keyword{internal}
\keyword{joyn's}
\keyword{locally}
\keyword{once}
\keyword{replaced}
\keyword{with}
\keyword{works}
