% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-joins.R
\name{unmatched_keys}
\alias{unmatched_keys}
\title{Check for unmatched keys}
\usage{
unmatched_keys(x, out, by)
}
\arguments{
\item{x}{input table to join}

\item{out}{output of join}

\item{by}{by argument, giving keys for join}
}
\value{
logical
}
\description{
Gives TRUE if unmatched keys, FALSE if not.
}
\keyword{internal}
